refactor: move organization-related views to module
Includes new `view_helpers.py` to prevent circular imports.
